refactor to use types from new userTypes file
