+++ plza-fe/src/redux/actions/userActions.js
@@ -1,31 +1,31 @@
 import API from "../../utils/API"
-import * as types from '../types'
+import * as types from '../types/userTypes'
 
 export const userLogin = ({ username, password }) => dispatch => {
-    dispatch({ type: types.LOGIN_START, payload: true })
+    dispatch({ type: types.USER_LOGIN_START, payload: true })
     API.post('/auth/user/login', { username, password })
         .then(res => {
             localStorage.setItem('token', JSON.stringify(res.data.token))
-            dispatch({ type: types.LOGIN_SUCCESS, payload: res.data.user })
+            dispatch({ type: types.USER_LOGIN_SUCCESS, payload: res.data.user })
             window.location.replace("/users/dash")
         })
         .catch(err => {
-            dispatch({ type: types.LOGIN_FAIL, payload: 'Invalid username or password' })
+            dispatch({ type: types.USER_LOGIN_FAIL, payload: 'Invalid username or password' })
         })
 }
 
 export const userRegister = (data) => dispatch => {
-    dispatch({ type: types.REGISTER_START, payload: true })
+    dispatch({ type: types.USER_REGISTER_START, payload: true })
     delete data.verify_password
     API.post('/auth/user/register', data)
         .then(res => {
             localStorage.setItem('token', JSON.stringify(res.data.token))
-            dispatch({ type: types.REGISTER_SUCCESS, payload: res.data.user })
+            dispatch({ type: types.USER_REGISTER_SUCCESS, payload: res.data.user })
             window.location.replace("/users/dash")
         })
         .catch(err => {
             dispatch({ 
-                type: types.REGISTER_FAIL, 
+                type: types.USER_REGISTER_FAIL, 
                 payload: {error: 'There was an error with your registration, please try again', isLoading: false} 
             })
         })
